feat(checkofficeid): Add new linter rule for OfficeID
This commit introduces a new linter rule in the `checkofficeid` package.
The rule checks if a struct with a `personal_data` security tag also
includes an `OfficeID` field of integer type. The linter rule is
integrated into the main function, replacing the previous `mylinter`
Analyzer. Unit tests for the new rule are also included.